Add debug facility
Some of the problems irqd sees are due to failed hardware, e. g. a CPU
core not running if it's broken.

Implementing workarounds in these cases is increasingly difficult, as
there currently is no way to replace files below /proc with "broken"
ones on the test system.

I therefore chose to add the possibility to replace selected files
(not necessarily below /proc) with custom ones.  To use a custom /proc
file simply start irqd with the IRQD_PREFIX variable set:

 $ IRQD_PREFIX=~/tmp/irqd-test irqd

Then place selected files in this directory, so e. g. /proc/stat
becomes ~/tmp/irqd-test/proc/stat.  It will then be used instead.

If the prefix is specified optional files below this directory
take precedence over the real files.

Of course this should be done with caution, as otherwise random input
in this file may lead to strange effects in irqd.  Though the parsers
should generally be good enough to not crash in this case, instead
they should just error out.

Signed-off-by: Holger Eitzenberger <holger@eitzenberger.org>