Initial configuration file
This is the first step into a general configurability.

If not specifying a config file all NICs are balanced over all CPUs,
which is usefull to keep compatility with older versions.  Also this
way there is still quite some improvement compared to not running
irqd at all.

If having a config file however all NICs to be balanced need to be
listed, other NICs are ignored.  A sample config file:

  cpuset "network" 0 4 {
  	devices {
  		interface "eth0";
  		interface "eth1";
  	};
  };

The above cpuset uses CPU0 to CPU3.  Multiple cpusets may be
specified, non-overlapping.  Each NIC can only be part of a
single cpuset.

Signed-off-by: Holger Eitzenberger <holger@eitzenberger.org>