evenly: improve distribution
Was not so evenly distributed.

And it was because of the fact that cpu_cmp() was sorting descending
instead of ascending.  Which was partly covered by the fact that in
add_queue() the ci_num_queues counter was incremented *after* insert.
Because of this the insert only case was ok.

Signed-off-by: Holger Eitzenberger <holger@eitzenberger.org>